+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.6.1] - 2024-05-15
+
+### Changed
+
+- Use a table writing package instead of the builtin package
+
+### Added
+
+- `gatecheck list --markdown` support for rendering markdown tables
+
 ## [0.6.0] - 2024-04-26
 
 ### Changed